Updated system to use global serial number
